Extract step injection logic into a separate step
